# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LMPS-API 1.0",
    "description": "This is a wip API for the LMPS and external ressources.",
    "termsOfService": "https://www.mac.de/en/terms-and-conditions/",
    "contact": {
      "email": "marius.kehl@mac.de"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://api.mac.de/api/v1",
      "description": "Production server"
    },
    {
      "url": "https://dev.api.mac.de/api/v1",
      "description": "Development server"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/projects": {
      "get": {
        "tags": [
          "project"
        ],
        "summary": "Get all projects",
        "description": "Returns all projects.\nThis endpoint is paginated.",
        "operationId": "getProjects",
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "name": "sort",
            "in": "query",
            "description": "The field to sort by.\nDefaults to id.",
            "required": false,
            "explode": true,
            "style": "deepObject",
            "schema": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "id",
                  "enum": [
                    "id",
                    "event_id",
                    "project_number",
                    "status",
                    "af_date",
                    "ag_date",
                    "au_date",
                    "customer_id",
                    "costunit_id",
                    "title",
                    "description",
                    "area",
                    "depth",
                    "hall_number",
                    "width",
                    "booth_number",
                    "booth_type",
                    "created_at",
                    "updated_at"
                  ],
                  "default": [
                    "id"
                  ]
                },
                "order": {
                  "type": "string",
                  "example": "asc",
                  "enum": [
                    "asc",
                    "desc"
                  ]
                }
              }
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "explode": true,
            "style": "deepObject",
            "schema": {
              "type": "object",
              "minProperties": 0,
              "properties": {
                "event_id": {
                  "type": "string",
                  "format": "uuid",
                  "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "project_number": {
                  "type": "string",
                  "example": 1234
                },
                "status": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "enum",
                    "example": "in progress",
                    "enum": [
                      "AN",
                      "AG",
                      "AF",
                      "OE",
                      "AU",
                      "XX"
                    ]
                  }
                },
                "af_date": {
                  "$ref": "#/components/schemas/DateFilter"
                },
                "ag_date": {
                  "$ref": "#/components/schemas/DateFilter"
                },
                "au_date": {
                  "$ref": "#/components/schemas/DateFilter"
                },
                "customer_id": {
                  "type": "string",
                  "format": "uuid",
                  "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "costunit_id": {
                  "type": "string",
                  "format": "uuid",
                  "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "title": {
                  "$ref": "#/components/schemas/StringFilter"
                },
                "description": {
                  "$ref": "#/components/schemas/StringFilter"
                },
                "booth_number": {
                  "$ref": "#/components/schemas/StringFilter"
                },
                "created_at": {
                  "$ref": "#/components/schemas/DateFilter"
                },
                "updated_at": {
                  "$ref": "#/components/schemas/DateFilter"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Project"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Meta"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorBadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "projects.read",
              "projects.read.all"
            ]
          },
          {
            "lmpsLegacy": []
          }
        ]
      },
      "post": {
        "tags": [
          "project"
        ],
        "summary": "Create a project",
        "description": "Creates a project.",
        "operationId": "createProject",
        "requestBody": {
          "description": "Project object\ndatabase_name: Projekt",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorBadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "projects.write",
              "projects.write.all"
            ]
          },
          {
            "lmpsLegacy": []
          }
        ]
      }
    },
    "/projects/{id}": {
      "get": {
        "tags": [
          "project"
        ],
        "summary": "Get a project",
        "description": "Returns a project.",
        "operationId": "getProject",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorBadRequest"
          },
          "404": {
            "$ref": "#/components/responses/ErrorNotFound"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "projects.read",
              "projects.read.all"
            ]
          },
          {
            "lmpsLegacy": []
          }
        ]
      },
      "put": {
        "tags": [
          "project"
        ],
        "summary": "Update a project",
        "description": "Updates a project.",
        "operationId": "updateProject",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "description": "Project object\ndatabase_name: Projekt",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorBadRequest"
          },
          "404": {
            "$ref": "#/components/responses/ErrorNotFound"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "projects.write",
              "projects.write.all"
            ]
          },
          {
            "lmpsLegacy": []
          }
        ]
      },
      "delete": {
        "tags": [
          "project"
        ],
        "summary": "Delete a project",
        "description": "Deletes a project.",
        "operationId": "deleteProject",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/ErrorBadRequest"
          },
          "404": {
            "$ref": "#/components/responses/ErrorNotFound"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "projects.delete"
            ]
          },
          {
            "lmpsLegacy": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Project": {
        "type": "object",
        "description": "Project object\ndatabase_name: Projekt",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "readOnly": true,
            "nullable": false,
            "maxLength": 36
          },
          "event_id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "maxLength": 36
          },
          "project_number": {
            "type": "string",
            "example": 1234,
            "maxLength": 3
          },
          "status": {
            "type": "string",
            "format": "enum",
            "example": "in progress",
            "maxLength": 2,
            "maxItems": 1,
            "enum": [
              "AN",
              "AG",
              "AF",
              "OE",
              "AU",
              "XX"
            ],
            "description": "AN = Request (Anfrage)\nAG = Offer (Angebot)\nAF = Order (Auftrag)\nOE = Open items (Offene Posten)\nAU = Completed (Abgeschlossen)\nXX = Canceled (Storniert)"
          },
          "af_date": {
            "type": "string",
            "format": "date",
            "example": "2020-01-01T00:00:00.000Z",
            "nullable": true
          },
          "ag_date": {
            "type": "string",
            "format": "date",
            "example": "2020-01-01T00:00:00.000Z",
            "nullable": true
          },
          "au_date": {
            "type": "string",
            "format": "date",
            "example": "2020-01-01T00:00:00.000Z",
            "nullable": true
          },
          "customer_id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "maxLength": 36
          },
          "costunit_id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "maxLength": 36
          },
          "title": {
            "type": "string",
            "example": "Test Project",
            "maxLength": 60
          },
          "description": {
            "type": "string",
            "example": "Test Project",
            "description": "This field can be pretty long as it is a \"clob\" in the database"
          },
          "area": {
            "type": "number",
            "format": "float",
            "example": 1234.56
          },
          "depth": {
            "type": "number",
            "format": "float",
            "example": 1234.56
          },
          "hall_number": {
            "type": "string",
            "example": 1234,
            "maxLength": 4
          },
          "width": {
            "type": "number",
            "format": "float",
            "example": 1234.56
          },
          "booth_number": {
            "type": "string",
            "example": 1234,
            "maxLength": 10
          },
          "booth_type": {
            "type": "string",
            "format": "enum",
            "maxLength": 5,
            "enum": [
              "KOPF",
              "REIHE",
              "ZELT",
              "BLOCK",
              "KEINE",
              "ECK",
              "PARTN"
            ],
            "description": "KOPF = Head-booth (Kopfstand)\nREIHE = Row-booth (Reihenstand)\nZELT = Tent-booth (Zeltstand)\nBLOCK = Block-booth (Blockstand)\nKEINE = not specified (keine Angabe)\nECK = Corner-booth (Eckstand)\nPARTN = Partner-booth (Partnerstand)",
            "example": "KOPF"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2020-01-01T00:00:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2020-01-01T00:00:00.000Z"
          }
        }
      },
...
      "Meta": {
        "type": "object",
        "properties": {
          "pagination": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "format": "int32",
                "example": 1,
                "description": "Total number of items available"
              },
              "count": {
                "type": "integer",
                "format": "int32",
                "example": 1,
                "description": "Number of items returned"
              },
              "per_page": {
                "type": "integer",
                "format": "int32",
                "example": 1,
                "description": "Number of items per page"
              },
              "current_page": {
                "type": "integer",
                "format": "int32",
                "example": 1,
                "description": "Current page"
              },
              "total_pages": {
                "type": "integer",
                "format": "int32",
                "example": 1,
                "description": "Total number of pages"
              },
              "links": {
                "type": "object",
                "properties": {
                  "previous": {
                    "type": "string",
                    "example": "http://example.com/api/v1/projects?page=1"
                  },
                  "next": {
                    "type": "string",
                    "example": "http://example.com/api/v1/projects?page=3"
                  }
                }
              }
            }
          }
        }
      },
      "StringFilter": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "enum",
            "enum": [
              "contains",
              "starts_with",
              "ends_with",
              "equals",
              "like"
            ],
            "example": "contains"
          },
          "value": {
            "type": "string",
            "example": "Test"
          }
        }
      },
      "DateFilter": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "enum",
            "enum": [
              "before",
              "after",
              "between"
            ]
          },
          "date": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time",
              "example": "2020-01-01T00:00:00.000Z"
            }
          }
        }
      }
    },
    "parameters": {
      "page": {
        "name": "page",
        "in": "query",
        "description": "The page number to return.\nDefaults to 1.",
        "required": false,
        "schema": {
          "type": "string",
          "format": "int32",
          "minimum": 1,
          "default": 1
        }
      },
      "limit": {
        "name": "limit",
        "in": "query",
        "description": "The number of items to return per page.\nDefaults to 10.",
        "required": false,
        "schema": {
          "type": "string",
          "format": "int32",
          "minimum": 1,
          "maximum": 100,
          "default": 10
        }
      },
      "id": {
        "name": "id",
        "in": "path",
        "description": "The id of the item to retrieve.",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "responses": {
      "ErrorNotFound": {
        "description": "The requested resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "format": "int32",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "The requested resource was not found"
                },
                "timestamp": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2020-01-01T00:00:00.000Z"
                },
                "path": {
                  "type": "string",
                  "example": "/api/v1/projects/123e4567-e89b-12d3-a456-426614174000"
                },
                "lmps-tracking-id": {
                  "type": "string",
                  "example": "123e4567-e89b-12d3-a456-426614174000"
                }
              }
            }
          }
        }
      },
      "ErrorBadRequest": {
        "description": "The request was invalid or cannot be otherwise served",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "format": "int32",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "The request was invalid or cannot be otherwise served"
                },
                "timestamp": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2020-01-01T00:00:00.000Z"
                },
                "path": {
                  "type": "string",
                  "example": "/api/v1/projects/123e4567-e89b-12d3-a456-426614174000"
                },
                "lmps-tracking-id": {
                  "type": "string",
                  "example": "123e4567-e89b-12d3-a456-426614174000"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/oauth/authorize",
            "tokenUrl": "/oauth/token",
            "refreshUrl": "/oauth/token",
            "scopes": {
              "project.read": "read access to projects a user is assigned to",
              "project.read.all": "read access to all projects",
              "project.write": "write access to projects a user is assigned to",
              "project.write.all": "write access to all projects",
              "project.delete": "delete access to all projects",
              "customer.read": "read access to customers a user is assigned to",
              "customer.read.all": "read access to all customers",
              "customer.write": "write access to customers a user is assigned to",
              "customer.write.all": "write access to all customers",
              "customer.delete": "delete access to all customers",
              "event.read": "read access to events a user is assigned to",
              "event.read.all": "read access to all events",
              "event.write": "write access to events a user is assigned to",
              "event.write.all": "write access to all events",
              "event.delete": "delete access to all events"
            }
          },
          "clientCredentials": {
            "tokenUrl": "/oauth/token",
            "refreshUrl": "/oauth/token",
            "scopes": {
              "project.read": "read access to projects",
              "project.write": "write access to projects",
              "project.delete": "delete access to projects",
              "customer.read": "read access to customers",
              "customer.write": "write access to customers",
              "customer.delete": "delete access to customers",
              "event.read": "read access to events",
              "event.write": "write access to events",
              "event.delete": "delete access to events"
            }
          }
...
  }
}